TST: Add tests for `MultiStairHandler`
This commit adds a new base class `_BaseTestMultiStairHandler`.
Adds rudimentary tests for `MultiStairHandler` using QUEST.
 `_BaseTestStairHandler` has been refactored to allow for more OOP
 usage.
